SahayakSetu Assistant/requirements.txt
fastapi
uvicorn[standard]
langchain==0.1.16
langchain-google-genai==1.0.3
python-dotenv